+++ internal/formatters/fmt_base.go
@@ -1,45 +1,45 @@
 package formatters
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"os"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
-	"unicode"
 
 	messages "github.com/cucumber/messages/go/v21"
 
 	"github.com/cucumber/godog/colors"
 	"github.com/cucumber/godog/formatters"
 	"github.com/cucumber/godog/internal/models"
+	"github.com/cucumber/godog/internal/snippets"
 	"github.com/cucumber/godog/internal/storage"
 	"github.com/cucumber/godog/internal/utils"
 )
 
 // BaseFormatterFunc implements the FormatterFunc for the base formatter.
-func BaseFormatterFunc(suite string, out io.Writer) formatters.Formatter {
-	return NewBase(suite, out)
+func BaseFormatterFunc(suite string, out io.Writer, snippetFunc string) formatters.Formatter {
+	return NewBase(suite, out, snippetFunc)
 }
 
 // NewBase creates a new base formatter.
-func NewBase(suite string, out io.Writer) *Base {
+func NewBase(suite string, out io.Writer, snippetFunc string) *Base {
 	return &Base{
-		suiteName: suite,
-		indent:    2,
-		out:       out,
-		Lock:      new(sync.Mutex),
+		snippetFunc: snippets.Find(snippetFunc),
+		suiteName:   suite,
+		indent:      2,
+		out:         out,
+		Lock:        new(sync.Mutex),
 	}
 }
 
 // Base is a base formatter.
 type Base struct {
-	suiteName string
-	out       io.Writer
-	indent    int
+	suiteName   string
+	out         io.Writer
+	indent      int
+	snippetFunc snippets.Func
 
 	Storage *storage.Storage
 	Lock    *sync.Mutex
@@ -185,7 +185,6 @@ func (f *Base) Summary() {
 		fmt.Fprintln(f.out, "")
 		fmt.Fprintln(f.out, "Randomized with seed:", colors.Yellow(seed))
 	}
-
 	if text := f.Snippets(); text != "" {
 		fmt.Fprintln(f.out, "")
 		fmt.Fprintln(f.out, yellow("You can implement step definitions for undefined steps with these snippets:"))
@@ -195,67 +194,5 @@ func (f *Base) Summary() {
 
 // Snippets returns code suggestions for undefined steps.
 func (f *Base) Snippets() string {
-	undefinedStepResults := f.Storage.MustGetPickleStepResultsByStatus(undefined)
-	if len(undefinedStepResults) == 0 {
-		return ""
-	}
-
-	var index int
-	var snips []undefinedSnippet
-	// build snippets
-	for _, u := range undefinedStepResults {
-		pickleStep := f.Storage.MustGetPickleStep(u.PickleStepID)
-
-		steps := []string{pickleStep.Text}
-		arg := pickleStep.Argument
-		if u.Def != nil {
-			steps = u.Def.Undefined
-			arg = nil
-		}
-		for _, step := range steps {
-			expr := snippetExprCleanup.ReplaceAllString(step, "\\$1")
-			expr = snippetNumbers.ReplaceAllString(expr, "(\\d+)")
-			expr = snippetExprQuoted.ReplaceAllString(expr, "$1\"([^\"]*)\"$2")
-			expr = "^" + strings.TrimSpace(expr) + "$"
-
-			name := snippetNumbers.ReplaceAllString(step, " ")
-			name = snippetExprQuoted.ReplaceAllString(name, " ")
-			name = strings.TrimSpace(snippetMethodName.ReplaceAllString(name, ""))
-			var words []string
-			for i, w := range strings.Split(name, " ") {
-				switch {
-				case i != 0:
-					w = strings.Title(w)
-				case len(w) > 0:
-					w = string(unicode.ToLower(rune(w[0]))) + w[1:]
-				}
-				words = append(words, w)
-			}
-			name = strings.Join(words, "")
-			if len(name) == 0 {
-				index++
-				name = fmt.Sprintf("StepDefinitioninition%d", index)
-			}
-
-			var found bool
-			for _, snip := range snips {
-				if snip.Expr == expr {
-					found = true
-					break
-				}
-			}
-			if !found {
-				snips = append(snips, undefinedSnippet{Method: name, Expr: expr, argument: arg})
-			}
-		}
-	}
-
-	sort.Sort(snippetSortByMethod(snips))
-
-	var buf bytes.Buffer
-	if err := undefinedSnippetsTpl.Execute(&buf, snips); err != nil {
-		panic(err)
-	}
-	// there may be trailing spaces
-	return strings.Replace(buf.String(), " \n", "\n", -1)
+	return f.snippetFunc(f.Storage)
 }